+++ jmix-localized-string-datatype/src/main/java/com/glebfox/jmix/locstr/validation/constraints/LocalizedStringSize.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2024 Gleb Gorelov.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.glebfox.jmix.locstr.validation.constraints;
+
+import com.glebfox.jmix.locstr.validation.constraints.impl.LocalizedStringSizeValidator;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Documented
+@Constraint(validatedBy = LocalizedStringSizeValidator.class)
+@Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE,
+        ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE})
+@Retention(RUNTIME)
+@Repeatable(LocalizedStringSize.List.class)
+public @interface LocalizedStringSize {
+
+    String message() default "{msg://com.glebfox.jmix.locstr/validation.constraints.LocalizedStringSize.message}";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+
+    int min() default 0;
+
+    int max() default Integer.MAX_VALUE;
+
+    @Documented
+    @Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE,
+            ElementType.CONSTRUCTOR, ElementType.PARAMETER, ElementType.TYPE_USE})
+    @Retention(RUNTIME)
+    @interface List {
+
+        LocalizedStringSize[] value();
+    }
+}